backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.9.2
